backend/requirements.txt
# backend/requirements.txt

# ウェブフレームワーク
fastapi>=0.95.0
uvicorn[standard]>=0.21.1

# データベース
sqlmodel>=0.0.8
sqlalchemy>=2.0.9

# HTTP通信
httpx>=0.24.0

# スクレイピング
beautifulsoup4>=4.12.0
lxml>=4.9.2

# ユーティリティ
python-dotenv>=1.0.0
loguru>=0.7.0
pydantic>=1.10.7

# 日付処理
python-dateutil>=2.8.2 
